+++ src/app/Balloon.App.Api/v2/Nodes.php
@@ -186,9 +186,7 @@ class Nodes extends Controller
                 $node = $node->setParent($parent, $conflict);
             }
 
-            if ($node->isDeleted()) {
-                $node->undelete($conflict);
-            }
+            $node->undelete($conflict);
 
             return [
                 'code' => 200,
@@ -969,6 +967,7 @@ class Nodes extends Controller
         foreach ($this->_getNodes($id, $path) as $node) {
             try {
                 $node->zip($archive);
+                //json_decode($stored, true),
             } catch (\Exception $e) {
                 $this->logger->debug('failed zip node in multi node request ['.$node->getId().']', [
                    'category' => get_class($this),